+++ GoCook_/app/build.gradle
@@ -34,6 +34,10 @@ dependencies {
     implementation 'androidx.appcompat:appcompat:1.4.1'
     implementation 'com.google.android.material:material:1.5.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.3'
+//    implementation 'com.google.firebase:firebase-database:20.0.5'
+    implementation 'com.google.firebase:firebase-database-ktx:20.0.5'
+
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     implementation 'com.google.firebase:firebase-database:20.0.5'
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
@@ -41,4 +45,9 @@ dependencies {
 
     // OCR 라이브러리 추가!
     implementation 'com.rmtheis:tess-two:9.1.0'
+
+    // 화면 맨 하단 바텀 네비게이션을 이용하기 위한 추가
+    implementation 'androidx.navigation:navigation-fragment:2.3.2'
+    implementation 'androidx.navigation:navigation-ui:2.3.2'
+
 }
