+++ GoCook_/app/google-services.json
@@ -1,0 +1,127 @@
+{
+  "project_info": {
+    "project_number": "437297234792",
+    "firebase_url": "https://realtimedb-355a8-default-rtdb.firebaseio.com",
+    "project_id": "realtimedb-355a8",
+    "storage_bucket": "realtimedb-355a8.appspot.com"
+  },
+  "client": [
+    {
+      "client_info": {
+        "mobilesdk_app_id": "1:437297234792:android:f8eeb22c6e9d5781094cdd",
+        "android_client_info": {
+          "package_name": "com.example.gocook_"
+        }
+      },
+      "oauth_client": [
+        {
+          "client_id": "437297234792-r3588ps497va8ci4asd24mi9idtghgp9.apps.googleusercontent.com",
+          "client_type": 3
+        }
+      ],
+      "api_key": [
+        {
+          "current_key": "AIzaSyBml0KeVi4Lol0tuE-bVLJU0rsDWNjru7c"
+        }
+      ],
+      "services": {
+        "appinvite_service": {
+          "other_platform_oauth_client": [
+            {
+              "client_id": "437297234792-r3588ps497va8ci4asd24mi9idtghgp9.apps.googleusercontent.com",
+              "client_type": 3
+            }
+          ]
+        }
+      }
+    },
+    {
+      "client_info": {
+        "mobilesdk_app_id": "1:437297234792:android:3a60db48d60818fa094cdd",
+        "android_client_info": {
+          "package_name": "com.example.realtimedb"
+        }
+      },
+      "oauth_client": [
+        {
+          "client_id": "437297234792-r3588ps497va8ci4asd24mi9idtghgp9.apps.googleusercontent.com",
+          "client_type": 3
+        }
+      ],
+      "api_key": [
+        {
+          "current_key": "AIzaSyBml0KeVi4Lol0tuE-bVLJU0rsDWNjru7c"
+        }
+      ],
+      "services": {
+        "appinvite_service": {
+          "other_platform_oauth_client": [
+            {
+              "client_id": "437297234792-r3588ps497va8ci4asd24mi9idtghgp9.apps.googleusercontent.com",
+              "client_type": 3
+            }
+          ]
+        }
+      }
+    },
+    {
+      "client_info": {
+        "mobilesdk_app_id": "1:437297234792:android:aeb43bccae33e390094cdd",
+        "android_client_info": {
+          "package_name": "com.example.test"
+        }
+      },
+      "oauth_client": [
+        {
+          "client_id": "437297234792-r3588ps497va8ci4asd24mi9idtghgp9.apps.googleusercontent.com",
+          "client_type": 3
+        }
+      ],
+      "api_key": [
+        {
+          "current_key": "AIzaSyBml0KeVi4Lol0tuE-bVLJU0rsDWNjru7c"
+        }
+      ],
+      "services": {
+        "appinvite_service": {
+          "other_platform_oauth_client": [
+            {
+              "client_id": "437297234792-r3588ps497va8ci4asd24mi9idtghgp9.apps.googleusercontent.com",
+              "client_type": 3
+            }
+          ]
+        }
+      }
+    },
+    {
+      "client_info": {
+        "mobilesdk_app_id": "1:437297234792:android:99ba31165dbdef7c094cdd",
+        "android_client_info": {
+          "package_name": "com.example.test2"
+        }
+      },
+      "oauth_client": [
+        {
+          "client_id": "437297234792-r3588ps497va8ci4asd24mi9idtghgp9.apps.googleusercontent.com",
+          "client_type": 3
+        }
+      ],
+      "api_key": [
+        {
+          "current_key": "AIzaSyBml0KeVi4Lol0tuE-bVLJU0rsDWNjru7c"
+        }
+      ],
+      "services": {
+        "appinvite_service": {
+          "other_platform_oauth_client": [
+            {
+              "client_id": "437297234792-r3588ps497va8ci4asd24mi9idtghgp9.apps.googleusercontent.com",
+              "client_type": 3
+            }
+          ]
+        }
+      }
+    }
+  ],
+  "configuration_version": "1"
+}
