+++ GoCook_/app/src/main/java/com/example/gocook_/JoinActivity.java
@@ -3,8 +3,6 @@ package com.example.gocook_;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
-import android.widget.Button;
-import android.widget.ImageButton;
 import android.widget.ImageView;
 
 import androidx.appcompat.app.AppCompatActivity;
@@ -21,7 +19,7 @@ public class JoinActivity extends AppCompatActivity {
         btn_back.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Intent intent = new Intent(getApplicationContext(), MainActivity.class);
+                Intent intent = new Intent(getApplicationContext(), MemberLoginActivity.class);
                 startActivity(intent); // 첫 화면 호출
             }
         });
