+++ GoCook_/app/src/main/res/layout/member_login.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:gravity="center"
     android:orientation="vertical"
-    tools:context=".MainActivity">
+    tools:context=".MemberLoginActivity">
 
     <LinearLayout
         android:layout_width="wrap_content"
