+++ GoCook_/app/src/main/res/menu/bottom_menu.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/item_home"
+        android:enabled="true"
+        android:icon="@drawable/ic_baseline_home_24"
+        android:title="홈화면"/>
+    <item
+        android:id="@+id/item_recipe"
+        android:enabled="true"
+        android:icon="@drawable/ic_baseline_article_24"
+        android:title="홈화면"/>
+    <item
+        android:id="@+id/item_setting"
+        android:enabled="true"
+        android:icon="@drawable/ic_baseline_menu_24"
+        android:title="홈화면"/>
+
+
+
+
+</menu>
